+++ src/ontology/bandit.hh
@@ -7,22 +7,22 @@ namespace ontology {
 
 class Exp3Light {
 public:
-  Exp3Light(int64_t arms, int64_t trials, double lossBound) : arms(arms), trials(trials), loss_bound(lossBound) {
+  Exp3Light(int64_t arms, int64_t trials, long double lossBound) : arms(arms), trials(trials), loss_bound(lossBound) {
     weight.resize(arms, 1);
-    weight_sum = static_cast<double>(arms);
+    weight_sum = static_cast<long double>(arms);
     expected_total_loss.resize(arms, 0);
 
     current_trial = 0;
     epoch = 0;
 
-    min_total_loss = std::numeric_limits<double>::max();
+    min_total_loss = std::numeric_limits<long double>::max();
     update_eta();
   }
 
 public:
   int64_t select_arm() {
-    double rand = absl::Uniform(gen, 0, weight_sum);
-    double rolling_sum = 0;
+    long double rand = absl::Uniform(gen, 0, weight_sum);
+    long double rolling_sum = 0;
 
     for (int64_t arm = 0; arm < arms; ++arm) {
       rolling_sum += weight[arm];
@@ -38,12 +38,12 @@ public:
 
   void update_weights(int64_t selected_arm, double loss) {
     // unbiased estimator for loss (loss of selection / probability of selection)
-    double loss_estimate = loss / (weight[selected_arm] / weight_sum);
+    long double loss_estimate = loss / (weight[selected_arm] / weight_sum);
 
     expected_total_loss[selected_arm] += loss_estimate;
     // update weight
-    double previous_weight = weight[selected_arm];
-    double new_weight = std::exp(-eta * expected_total_loss[selected_arm] / loss_bound);
+    long double previous_weight = weight[selected_arm];
+    long double new_weight = std::exp(-eta * expected_total_loss[selected_arm] / loss_bound);
     weight[selected_arm] = new_weight;
     weight_sum += new_weight - previous_weight;
 
@@ -57,42 +57,42 @@ public:
     }
   }
 
-  void restart_with_loss_bound(double new_lb) {
+  void restart_with_loss_bound(long double new_lb) {
     loss_bound = new_lb;
     std::fill(weight.begin(), weight.end(), 1);
-    weight_sum = static_cast<double>(arms);
+    weight_sum = static_cast<long double>(arms);
     std::fill(expected_total_loss.begin(), expected_total_loss.end(), 0);
-    min_total_loss = std::numeric_limits<double>::max();
+    min_total_loss = std::numeric_limits<long double>::max();
 
     epoch = 0;
     trials = trials - current_trial;
     current_trial = 0;
 
-    min_total_loss = std::numeric_limits<double>::max();
+    min_total_loss = std::numeric_limits<long double>::max();
     update_eta();
   }
 
-  double get_normalized_weight(int64_t arm) {
+  long double get_normalized_weight(int64_t arm) {
     return weight[arm] / weight_sum;
   }
 
 private:
   void update_eta() {
-    eta = std::sqrt((2 * (std::log(arms) + static_cast<double>(arms) * std::log(trials))) /
-                    (static_cast<double>(arms) * std::pow(4, epoch)));
+    eta = std::sqrt((2 * (std::log(arms) + static_cast<long double>(arms) * std::log(trials))) /
+                    (static_cast<long double>(arms) * std::pow(4, epoch)));
   }
 
 private:
   int64_t arms;
   int64_t trials;
   int64_t current_trial;
-  double loss_bound;
+  long double loss_bound;
   int64_t epoch;
-  double eta;
-  std::vector<double> weight;
-  double weight_sum;
-  std::vector<double> expected_total_loss;
-  double min_total_loss;
+  long double eta;
+  std::vector<long double> weight;
+  long double weight_sum;
+  std::vector<long double> expected_total_loss;
+  long double min_total_loss;
   absl::BitGen gen;
 };
 
@@ -106,7 +106,7 @@ public:
     return bandit_solver.select_arm();
   }
 
-  void update_weights(int64_t selected_arm, double loss) {
+  void update_weights(int64_t selected_arm, long double loss) {
     if (loss > loss_bound) {
       int64_t u = std::ceil(std::log2(loss));
       loss_bound = std::pow(2, u);
@@ -116,7 +116,7 @@ public:
     }
   }
 
-  double get_normalized_weight(int64_t arm) { return bandit_solver.get_normalized_weight(arm);
+  long double get_normalized_weight(int64_t arm) { return bandit_solver.get_normalized_weight(arm);
   }
 
 private:
@@ -125,7 +125,7 @@ private:
   int64_t current_trial;
 
   int64_t epoch;
-  double loss_bound;
+  long double loss_bound;
   Exp3Light bandit_solver;
 };
